Function parent registration request
Thêm function cho parent đăng kí tài khoản

+++ backend/src/main/java/com/swp391_8/schoolhealth/dto/StudentDTO.java
@@ -28,7 +28,7 @@ public class StudentDTO {
         this.className = student.getClassName();
         if (student.getUser() != null) {
             this.userUsername = student.getUser().getUsername();
-            this.userId = student.getUser().getUserId();
+            this.userId = student.getUser().getUserId().intValue();
         }
     }
 }
